styling and ESlint error fix
