+++ public/app/plugins/panel/barchart/BarChartPanel.tsx
@@ -50,7 +50,7 @@ export const BarChartPanel = (props: PanelProps<Options>) => {
     markers,
   } = options;
 
-  const markerGroups = markers.markerGroups || [];
+  const { markerGroups } = markers;
   // size-dependent, calculated opts that should cause viz re-config
   let { orientation, xTickLabelMaxLength = 0 } = options;
 
@@ -92,7 +92,8 @@ export const BarChartPanel = (props: PanelProps<Options>) => {
   const totalSeries = Math.max(0, (info.series[0]?.fields.length ?? 0) - 1);
 
   const preparedMarkers = useMemo(
-    () => prepMarkers(vizSeries[0]?.fields ?? [], markerData ?? [], options.markers.markerGroups ?? [], stacking, theme),
+    () =>
+      prepMarkers(vizSeries[0]?.fields ?? [], markerData ?? [], options.markers.markerGroups ?? [], stacking, theme),
     [markerData, options.markers.markerGroups, stacking, vizSeries, theme]
   );
