+++ public/app/plugins/panel/barchart/bars.ts
@@ -714,8 +714,9 @@ export function populateMarkerList(
     if (
       marker.groupIdx !== dataIdx ||
       seriesIdx !== marker.seriesIdx //match to bar
-    )
+    ) {
       continue;
+    }
 
     if (xOri === ScaleOrientation.Horizontal) {
       // Compute marker position at center of bar
@@ -731,7 +732,7 @@ export function populateMarkerList(
       const m: MarkerDrawingArgs = {
         x: markerX!,
         y: resolvedY,
-        opts: { ...marker.opts, size: marker.opts.size/2 },
+        opts: { ...marker.opts, size: marker.opts.size / 2 },
         isRotated: false,
       };
       resolvedMarkerList.push(m);
@@ -748,7 +749,7 @@ export function populateMarkerList(
       const m: MarkerDrawingArgs = {
         x: resolvedX!,
         y: markerY!,
-        opts: { ...marker.opts, size: marker.opts.size/2 },
+        opts: { ...marker.opts, size: marker.opts.size / 2 },
         isRotated: true,
       };
       resolvedMarkerList.push(m);
