+++ public/app/plugins/panel/barchart/utils.ts
@@ -9,7 +9,6 @@ import {
   GrafanaTheme2,
   PanelData,
   cacheFieldDisplayNames,
-  fieldColorModeRegistry,
   formattedValueToString,
   getDisplayProcessor,
   getFieldColorModeForField,
@@ -46,7 +45,6 @@ import { BarsOptions, getConfig } from './bars';
 import { PreparedMarker, MarkerGroup } from './markerTypes';
 import { FieldConfig, Options, defaultFieldConfig } from './panelcfg.gen';
 // import { isLegendOrdered } from './utils';
-import color from 'tinycolor2';
 
 interface BarSeries {
   series: DataFrame[];
@@ -620,16 +618,11 @@ export function prepMarkers(
 
     const fi = markerFields[i];
 
-    
-    let fieldColor = theme.visualization.getColorByName(m.opts.color)
+    let fieldColor = theme.visualization.getColorByName(m.opts.color);
     //Color override
-    if(fi.config.custom?.color) fieldColor = theme.visualization.getColorByName(fi.config.custom.color)
-
-    
-
-    // const colorMode = getFieldColorModeForField(fi);
-    // const scaleColor = getFieldSeriesColor(fi, theme);
-    // var seriesColor = scaleColor.color;
+    if (fi.config.custom?.color) {
+      fieldColor = theme.visualization.getColorByName(fi.config.custom.color);
+    }
 
     const targetIdx = vizFields.findIndex((f) => f.name === m.targetField);
 
@@ -641,7 +634,7 @@ export function prepMarkers(
             yValue: fi.values[j],
             seriesIdx: targetIdx,
             yScaleKey: fi.config.unit || FIXED_UNIT,
-            opts: {...m.opts, color: fieldColor},
+            opts: { ...m.opts, color: fieldColor },
           };
 
           prepMarkerList.push(pm);
@@ -659,7 +652,7 @@ export function prepMarkers(
             yValue: yTotal + fi.values[j],
             seriesIdx: targetIdx,
             yScaleKey: fi.config.unit || FIXED_UNIT,
-            opts: {...m.opts, color: fieldColor},
+            opts: { ...m.opts, color: fieldColor },
           };
 
           prepMarkerList.push(pm);
@@ -685,7 +678,7 @@ export function prepMarkers(
             yValue: val === 0 ? 0 : val / yTotal,
             seriesIdx: targetIdx,
             yScaleKey: fi.config.unit || FIXED_UNIT,
-            opts: {...m.opts, color: fieldColor},
+            opts: { ...m.opts, color: fieldColor },
           };
 
           prepMarkerList.push(pm);
